+++ packages/react-liveness/src/components/FaceLivenessDetector/shared/Hint.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Flex, Loader, View } from '@aws-amplify/ui-react';
+import { View } from '@aws-amplify/ui-react';
 
 import { IlluminationState, FaceMatchState } from '../service';
 
@@ -11,7 +11,7 @@ import {
 } from '../hooks';
 import { Toast } from './Toast';
 import { HintDisplayText } from '../displayText';
-import { LivenessClassNames } from '../types/classNames';
+import { ToastWithLoader } from './ToastWithLoader';
 
 export const selectErrorState = createLivenessSelector(
   (state) => state.context.errorState
@@ -33,6 +33,16 @@ export const selectFaceMatchStateBeforeStart = createLivenessSelector(
 export interface HintProps {
   hintDisplayText: Required<HintDisplayText>;
 }
+
+const defaultToast = (text: string, isInitial = false) => {
+  return (
+    <Toast size="large" variation="primary" isInitial={isInitial}>
+      <View aria-live="polite" aria-label={text}>
+        {text}
+      </View>
+    </Toast>
+  );
+};
 
 export const Hint: React.FC<HintProps> = ({ hintDisplayText }) => {
   const [state] = useLivenessActor();
@@ -80,11 +90,7 @@ export const Hint: React.FC<HintProps> = ({ hintDisplayText }) => {
 
   const getInstructionContent = () => {
     if (isStartView) {
-      return (
-        <Toast size="large" variation="primary" isInitial>
-          {hintDisplayText.hintCenterFaceText}
-        </Toast>
-      );
+      return defaultToast(hintDisplayText.hintCenterFaceText, true);
     }
 
     if (errorState ?? (isCheckFailed || isCheckSuccessful)) {
@@ -94,17 +100,11 @@ export const Hint: React.FC<HintProps> = ({ hintDisplayText }) => {
     if (!isRecording) {
       if (isCheckFaceDetectedBeforeStart) {
         if (faceMatchStateBeforeStart === FaceMatchState.TOO_MANY) {
-          return (
-            <Toast size="large" variation="primary">
-              {FaceMatchStateStringMap[faceMatchStateBeforeStart]}
-            </Toast>
+          return defaultToast(
+            FaceMatchStateStringMap[faceMatchStateBeforeStart]!
           );
         }
-        return (
-          <Toast size="large" variation="primary">
-            {hintDisplayText.hintMoveFaceFrontOfCameraText}
-          </Toast>
-        );
+        return defaultToast(hintDisplayText.hintMoveFaceFrontOfCameraText);
       }
 
       // Specifically checking for false here because initially the value is undefined and we do not want to show the instruction
@@ -112,50 +112,28 @@ export const Hint: React.FC<HintProps> = ({ hintDisplayText }) => {
         isCheckFaceDistanceBeforeRecording &&
         isFaceFarEnoughBeforeRecordingState === false
       ) {
-        return (
-          <Toast size="large" variation="primary">
-            {hintDisplayText.hintTooCloseText}
-          </Toast>
-        );
+        return defaultToast(hintDisplayText.hintTooCloseText);
       }
 
       if (isNotRecording) {
         return (
-          <Toast>
-            <Flex className={LivenessClassNames.HintText}>
-              <Loader />
-              <View>{hintDisplayText.hintConnectingText}</View>
-            </Flex>
-          </Toast>
+          <ToastWithLoader displayText={hintDisplayText.hintConnectingText} />
         );
       }
 
       if (isUploading) {
         return (
-          <Toast>
-            <Flex className={LivenessClassNames.HintText}>
-              <Loader />
-              <View>{hintDisplayText.hintVerifyingText}</View>
-            </Flex>
-          </Toast>
+          <ToastWithLoader displayText={hintDisplayText.hintVerifyingText} />
         );
       }
 
       if (illuminationState && illuminationState !== IlluminationState.NORMAL) {
-        return (
-          <Toast size="large" variation="primary">
-            {IlluminationStateStringMap[illuminationState]}
-          </Toast>
-        );
+        return defaultToast(IlluminationStateStringMap[illuminationState]);
       }
     }
 
     if (isFlashingFreshness) {
-      return (
-        <Toast size="large" variation="primary">
-          {hintDisplayText.hintHoldFaceForFreshnessText}
-        </Toast>
-      );
+      return defaultToast(hintDisplayText.hintHoldFaceForFreshnessText);
     }
 
     if (isRecording && !isFlashingFreshness) {
@@ -178,7 +156,9 @@ export const Hint: React.FC<HintProps> = ({ hintDisplayText }) => {
             faceMatchState === FaceMatchState.TOO_CLOSE ? 'error' : 'primary'
           }
         >
-          {resultHintString}
+          <View aria-live="polite" aria-label={resultHintString}>
+            {resultHintString}
+          </View>
         </Toast>
       );
     }