+++ packages/react-auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/ui-react-auth
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`c526cf53b`](https://github.com/aws-amplify/amplify-ui/commit/c526cf53bc07bb85c0789aa5edfdfcb4485fa5f5), [`cc2740b9a`](https://github.com/aws-amplify/amplify-ui/commit/cc2740b9a033e587715a37c379166388b267ff4e)]:
+  - @aws-amplify/ui-react@6.0.5
+  - @aws-amplify/ui@6.0.4
+  - @aws-amplify/ui-react-core@3.0.4
+  - @aws-amplify/ui-react-core-auth@0.0.6
+
 ## 0.0.6
 
 ### Patch Changes